refactor: added draggable to tracklist and deleted from tracks
